Move helper into UtxoView
